Extract C and C++ implementations into their respective source files
